Compute t-statistic map with or without Empirical Bayes.

Compute t-statistic map with or without Empirical Bayes.

getTMap(
  V,
  design,
  columnIndex = NULL,
  weights = NULL,
  brain.mask,
  eBayes = TRUE,
  output.file = "Tmap.nii.gz",
  returnObject = TRUE,
  writeToDisk = FALSE,
  verbose = TRUE,
  ...
)

Arguments

V

Matrix of voxels intensities (rows are voxels, columns are subjects).

design

Design matrix for the covariates of interest, with rows corresponding to subjects and columns to coefficients to be estimated.

columnIndex

Numeric indicating which covariate of the design matrix should be considered.

weights

Non-negative observation weights. Can be a numeric matrix of individual weivhts, of same size as the voxel matrix, or a numeric vector of subject weights with length equal to 'ncol' of the voxel matrix, or a nmeric vector of voxel weights with length equal to 'nrow' of the voxel matrix.

brain.mask

Path of a NIfTI file mask to specify subset of voxels for which the t-statistics should be computed.

eBayes

Should the limma Empirical Bayes method be used to compute moderated t-statistics?

output.file

Filename of the NIfTI file to be saved to disk, containing the computed t-statistics.

returnObject

Should the t-statistics be returned as a matrix?

writeToDisk

Should the t-statistics image be saved to the disk?

verbose

Should messages be printed?

...

additional arguments to pass to lmFit

Value

If returnObject is TRUE, a matrix of the t-statitics is returned, and if writeToDisk is TRUE, the t-statistics image is saved to disk as a NIfTI file.

Author

Jean-Philippe Fortin