Intersection of a list of brain masks

Intersection of a list of brain masks.

maskIntersect(list, output.file = NULL, prob = 1, reorient = FALSE,
  returnObject = TRUE, writeToDisk = TRUE, verbose = TRUE)

Arguments

list
List of the mask input files (nifti files) or mask nifti objects.

output.file
Optional filename for the output image.

prob
Probability of intersection. Ex: 0.9 will return a mask for voxels that intersect 90% of the time.

reorient
Should the input images be reoriented? FALSE by default.

returnObject
Should a nifti object be returned?

writeToDisk
Should the intersection mask be saved to the disk?

verbose
Should messages be printed?

Value

If returnObject is TRUE, a nifti object is returned, and if writeToDisk is TRUE, the intersection mask is saved to disk as a NIfTI file.

Examples

if (require(EveTemplate)){
  path1 <- path2 <- getEvePath()
  inter <- maskIntersect(list(path1,path2), writeToDisk=FALSE)
}

#> Loading required package: EveTemplate

#> Loading required package: oro.nifti

#> oro.nifti 0.10.2