Histogram matching intensity normalization.

Histogram mathing intensity normalization.

normalizeHM(
  input.files,
  output.files = NULL,
  brain.mask = NULL,
  type = c("T1", "T2", "FLAIR", "PD"),
  writeToDisk = FALSE,
  returnMatrix = TRUE,
  verbose = TRUE
)

Arguments

input.files

Vector of filenames for the input images. Must be NIfTI files.

output.files

Optional vector of filenames for the output images. By default, will be the input.files with suffix "HM".

brain.mask

Filename for the brain binary mask specifying the template space brain. Must be a NIfTI file.

type

What modality is used? Should be one of T1, T2, FLAIR or PD.

writeToDisk

Should the scans be saved to the disk? FALSE by default.

returnMatrix

Should the matrix of intensities be returned? FALSE by default.

verbose

Should messages be printed?

Value

if returnMatrix is FALSE, no value returned, but Histogram-matching-normalized images are saved. If returnMatrix is TRUE, Histogram-matching-normalized images are saved and a matrix of normalized intensities is returned.

Author

Jean-Philippe Fortin